improve prefix: use >= and <, instead of LIKE.
