+++ util/rich-indexer/Cargo.toml
@@ -20,6 +20,7 @@ ckb-notify = { path = "../../notify", version = "= 0.115.0-pre" }
 ckb-types = { path = "../types", version = "= 0.115.0-pre" }
 futures = "0.3"
 log = "0.4"
+num-bigint = "0.4"
 once_cell = "1.8.0"
 sql-builder = "3.1"
 sqlx = { version = "0.6", features = ["runtime-tokio-native-tls", "any", "sqlite", "postgres"] }
